+++ src/algorithms/Scale.php
@@ -13,7 +13,7 @@ use yii\base\InvalidConfigException;
  * Scale image algorithm.
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
- * @since 1.1
+ * @since 2.0
  */
 class Scale extends AbstractAlgorithm
 {