+++ tests/unit/helpers/TinifyDataTest.php
@@ -16,7 +16,7 @@ use PHPUnit\Framework\Constraint\IsType;
  * Test case for TinifyData helper.
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
- * @since 1.1
+ * @since 2.0
  */
 class TinifyDataTest extends TestCase
 {